In game GUI token counter, logic for enemy hit and target num of tokens reached
The GUI should be easily adapted to whatever else we want to add to it. Detecting when any enemy is hit, as well as whenever TARGET_TOKENS is reached.  Also, added 4 different directions that the linear enemies can move instead of just always going down and to the right at the start.