Implemented the winning and losing conditions
Created a new game end pane. Implemented code to make the tokens or enemy collision system trigger a response.
Ran into issues with parent node exceptions.
Currently, a large game breaking bug is in the game. More specifically , when a game is won or lost and a new one is started in the same instance, the game creates multiple instances.