End Game Fixes
Small edits to fix previous bugs. Paused game loop anytime our token goal is reached or an enemy is hit. Changed player to a public class and instantiated it using the constructor rather than get instance, I know this isn't how we wanted to do but thought I would try and it works.